requirements.txt
Flask
python-telegram-bot
langchain
sentence-transformers
langchain-community
langchain-huggingface
langchain-groq
langchain-core